+++ packages/web/website/.gitignore
@@ -33,6 +33,3 @@ yarn-error.log*
 # vercel
 .vercel
 node_modules
-
-# authentication information
-*.json
